remove code coverage for js dev files
I wanted to unit test private functions but it turned out to be
more pain than gain. As I splitted the code into several small
development files, the dependencies and paths started to become
difficult to manage.

It wasn't a good idea from the start so I think it's best to rollback.